+++ src/Command/Analyze.php
@@ -87,9 +87,10 @@ class Analyze implements Command
         // ext dir
         $config = $this->controller->getComposerConfig();
 
-        /** @phpstan-ignore-next-line */
         if (
+            /** @phpstan-ignore-next-line */
             !isset($config['require-dev']['decodelabs/phpstan-decodelabs']) &&
+            /** @phpstan-ignore-next-line */
             !isset($config['require']['decodelabs/phpstan-decodelabs'])
         ) {
             $this->controller->run('composer', 'require', 'decodelabs/phpstan-decodelabs', '--dev');