+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed check for phpstan-decodelabs
+
 ## v0.1.13 (2022-11-21)
 * Fixed local specialised phpstan script calls