+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Fixed check for phpstan-decodelabs
 * Added null check in template config lookups
+* Reload composer config after init-package
 
 ## v0.1.13 (2022-11-21)
 * Fixed local specialised phpstan script calls