+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Init repo before init-package
 * Reload composer config after init-package
 * Generate CI workflow after init-package
+* Replaced template comment pattern
 
 ## v0.1.13 (2022-11-21)
 * Fixed local specialised phpstan script calls